Rename examples_level_zero.* to examples_level_zero_helpers.*
