+++ examples/common/examples_level_zero_helpers.h
