+++ examples/level_zero_shared_memory/CMakeLists.txt
@@ -45,8 +45,9 @@ message(STATUS "Level Zero include directory: ${LEVEL_ZERO_INCLUDE_DIRS}")
 
 # build the example
 set(EXAMPLE_NAME umf_example_level_zero_shared_memory)
-add_executable(${EXAMPLE_NAME} level_zero_shared_memory.c
-                               ${UMF_EXAMPLE_DIR}/common/examples_level_zero.c)
+add_executable(
+    ${EXAMPLE_NAME} level_zero_shared_memory.c
+                    ${UMF_EXAMPLE_DIR}/common/examples_level_zero_helpers.c)
 target_include_directories(${EXAMPLE_NAME} PRIVATE ${LIBUMF_INCLUDE_DIRS}
                                                    ${UMF_EXAMPLE_DIR}/common)
 target_link_directories(${EXAMPLE_NAME} PRIVATE ${LIBUMF_LIBRARY_DIRS}
